refactor: 패키지 이동 및 서비스명 변경
