+++ core/data/src/main/java/com/saeongjima/data/api/TokenService.kt
@@ -1,9 +1,11 @@
-package com.saeongjima.data.di
+package com.saeongjima.data.api
 
+import com.saeongjima.data.api.model.SignInRequest
+import retrofit2.Response
 import retrofit2.http.Body
 import retrofit2.http.POST
 
-interface SignInService {
+interface TokenService {
     @POST("/login")
     suspend fun postSignIn(
         @Body
